Added some player input to move the camera a bit

+++ AquilaEngine/include/Timer.h
@@ -8,7 +8,7 @@ struct BenchmarkInfo {
 	//std::chrono::time<long long, std::nano> time;
 };
 
-double Bench_GetMiliseconds(const BenchmarkInfo & bench)
+float Bench_GetMiliseconds(const BenchmarkInfo & bench)
 {
 	//auto diff = std::chrono::duration_cast<std::chrono::nanoseconds>(bench.time);
 	return bench.time;//diff.count();
@@ -21,5 +21,5 @@ void Bench_End(BenchmarkInfo & bench)
 {
 	auto now = std::chrono::high_resolution_clock::now();
 
-	bench.time = std::chrono::duration_cast<std::chrono::nanoseconds>(now - bench.stopwatch).count() * 0.000001;
+	bench.time = std::chrono::duration_cast<std::chrono::nanoseconds>(now - bench.stopwatch).count() * 0.000001f;
 }
